requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.0
mccabe==0.7.0
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
sqlparse==0.5.0
